requirements.txt
# PolyX Telegram Bot
python-telegram-bot>=20.0
aiosqlite>=0.19.0
cryptography>=41.0.0
aiohttp>=3.9.0
requests
py-clob-client
python-dotenv>=1.0.0
eth-account>=0.9.0
web3>=6.0.0
PyJWT>=2.8.0
stripe>=8.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
